requirements.txt
opencv-python
pillow
numpy
pytesseract
easyocr
torch
python-Levenshtein
streamlit
scikit-learn
matplotlib
tqdm
